Activity Log: Replace API source for credentials display in settings
Previously we were hitting the deprecated endpoint at
`/activity-log/{ site }/rewind` to get information about whether or not
to show the Rewind credentials form in the **Security** settings.
This relied on client-side logic to determine if Rewind is `active`
or not but we have consolidated this logic in the server through
the `/sites/{ site }/rewind` API.

In this patch we're simply replacing the data source to use the
standardized API to ensure consistency in the display of the form.