update require --> import where applicable.
Background info:

in es6-land and webpack, if you want to require something that is exported in an es6 fashion, you need to write this:
require( 'module' ).default.

The default is not assumed.  Named imports work as expected and you write
require( 'module' ).namedExport.  So in a sense, default is just another name.

This PR is about converting as many requires into imports as possible because that makes everything less confusing and increases the scope of what webpack can statically analyze.

The only situations in which it is advisable to let a require stick around, is when it is only conditionally imported based on NODE_ENV.