Fix eslint errors regarding my-sites/sidebar class names.
For some reason these errors got picked up after rebasing this
branch to solve a conflict. Perhaps a rule has changed but it
doesn't appear to be affecting other builds.

The only other place a .stats class is used is in the
appbanner https://git.io/vN9VI but isn't affected by this
change.

Similarly the menu-link-text class doesn't have any CSS
rules associated with it and so is safe to prefix.